requirements.txt
Django==1.9.6
tornado==4.0.2
future==0.16.0
psutil==5.4.0
simplejson==3.11.1
paramiko==2.3.1
django-reversion==2.0.10
xadmin==0.6.1
requests==2.18.2
dwebsocket=0.4.2
django-celery==3.2.1
celery==3.1.25
